[refactoring]review로직 detail페이지에서 분리
