+++ src/pages/seulhaewon/Detail/Detail.scss
@@ -121,26 +121,6 @@ p{
     background-color: #EFEFEF;
 }
 
-.review_id{
-    width:70px;
-    padding:10px;
-    background-color: #EFEFEF;
-    border: 0px;
-    margin-bottom:10px;
-    margin-right:10px;
-}
-.review{
-    width:360px;
-    padding:10px;
-    background-color: #EFEFEF;
-    border: 0px;
-    margin-bottom:10px;
-    margin-right:10px;
-}
-.reviewBox{
-    display:flex;
-}
-
 footer{
     margin-top:100px;
     min-height:20vh;
@@ -173,15 +153,6 @@ h4{
     padding:0px;
     margin:10px;
     font-weight: bold;
-}
-
-.reviewContainer{
-    margin-bottom:10px;
-}
-
-.reviewBox{
-    display:flex;
-    justify-content: space-between;
 }
 
 .iconBox{
