+++ src/pages/seulhaewon/review/review.scss
@@ -1,0 +1,31 @@
+* {
+    box-sizing: border-box;
+}
+.reviewContainer{
+    margin-bottom:10px;
+}
+
+.reviewBox{
+    display:flex;
+    justify-content: space-between;
+}
+
+.review_id{
+    width:70px;
+    padding:10px;
+    background-color: #EFEFEF;
+    border: 0px;
+    margin-bottom:10px;
+    margin-right:10px;
+}
+.review{
+    width:360px;
+    padding:10px;
+    background-color: #EFEFEF;
+    border: 0px;
+    margin-bottom:10px;
+    margin-right:10px;
+}
+b{
+    font-weight:bold;
+}
